azure storage versions busted, GCS storage working
