+++ .env_sample
@@ -9,6 +9,12 @@ DB_PORT=5432
 RABBITMQ_DEFAULT_USER=rabbit-username
 RABBITMQ_DEFAULT_PASS=rabbit-password-you-should-change
 
+# -----------------------------------------------------------------------------
+# Storage
+#
+# Pick a storage from below and uncomment that section
+# -----------------------------------------------------------------------------
+
 # # S3 storage example
 # DEFAULT_FILE_STORAGE=storages.backends.s3boto3.S3Boto3Storage
 # AWS_ACCESS_KEY_ID=12312312312312312331223
@@ -19,14 +25,27 @@ RABBITMQ_DEFAULT_PASS=rabbit-password-you-should-change
 # AWS_QUERYSTRING_AUTH=False
 
 
-# Minio local storage example
-MINIO_ACCESS_KEY=testkey
-MINIO_SECRET_KEY=testsecret
+# # Minio local storage example
+# MINIO_ACCESS_KEY=testkey
+# MINIO_SECRET_KEY=testsecret
+#
+# DEFAULT_FILE_STORAGE=storages.backends.s3boto3.S3Boto3Storage
+# AWS_ACCESS_KEY_ID=testkey
+# AWS_SECRET_ACCESS_KEY=testsecret
+# AWS_STORAGE_BUCKET_NAME=public
+# AWS_STORAGE_PRIVATE_BUCKET_NAME=private
+# AWS_S3_ENDPOINT_URL=http://docker.for.mac.localhost:9000/
+# AWS_QUERYSTRING_AUTH=False
 
-DEFAULT_FILE_STORAGE=storages.backends.s3boto3.S3Boto3Storage
-AWS_ACCESS_KEY_ID=testkey
-AWS_SECRET_ACCESS_KEY=testsecret
-AWS_STORAGE_BUCKET_NAME=public
-AWS_STORAGE_PRIVATE_BUCKET_NAME=private
-AWS_S3_ENDPOINT_URL=http://docker.for.mac.localhost:9000/
-AWS_QUERYSTRING_AUTH=False
+# # Azure
+# DEFAULT_FILE_STORAGE=storages.backends.azure_storage.AzureStorage
+# AZURE_ACCOUNT_NAME=storage-account-name
+# AZURE_ACCOUNT_KEY=the-key==
+# AZURE_CONTAINER=public
+# BUNDLE_AZURE_CONTAINER=private
+
+# # GCS
+# DEFAULT_FILE_STORAGE=storages.backends.gcloud.GoogleCloudStorage
+# GS_PUBLIC_BUCKET_NAME=public
+# GS_PRIVATE_BUCKET_NAME=private
+# GOOGLE_APPLICATION_CREDENTIALS=/app/certs/google-storage-api.json
