+++ .gitignore
@@ -9,6 +9,7 @@ staticfiles/
 uploads/
 venv/
 var/
+certs/
 
 src/static/output.css
 src/static/output.js
