+++ docker-compose.yml
@@ -111,7 +111,6 @@ services:
   site_worker:
     # This auto-reloads
     command: bash -c "watchmedo auto-restart -p '*.py' --recursive -- celery -A comp_worker worker"
-    env_file: .env
     working_dir: /app/src
     build:
       context: .
