+++ requirements.txt
@@ -1,31 +1,45 @@
 Django==2.1
-psycopg2==2.7.3
-whitenoise==3.3.0
-coreapi==2.3.3
-djangorestframework==3.8.2
-markdown==2.6.11
-pygments==2.2.0
-waitress==1.0.2
-django-filter==2.0.0
 django-oauth-toolkit==1.0.0
 django-cors-middleware==1.3.1
-django-storages==1.6.6
-dj-database-url==0.4.2
 social-auth-core==1.5.0
 social-auth-app-django==2.0.0
 django-extensions==1.9.0
 channels==1.1.8
 asgi_redis==1.4.3
-drf-writable-nested==0.4.1
 django-extra-fields==0.9
 pillow==5.0.0
 celery==4.2.1
 gunicorn==19.9.0
 flower==0.9.2
 gevent==1.3.5
-gevent-eventemitter==2.0
-greenlet==0.4.14
 pyyaml==3.13
 watchdog==0.8.3
 python-dateutil==2.7.3
+
+# Database
+dj-database-url==0.4.2
+psycopg2==2.7.3
+
+# Storage
+django-storages[google]==1.6.6
+
+#git+https://github.com/jschneier/django-storages.git@master
+#-e 'git+https://github.com/mikebz/django-storages.git#egg=upgrade_azure'
+#azure==2.0.0
+#azure-storage==0.34.3
+
+
+#azure-storage-blob==1.3.1
+#azure-storage<=0.20.0
+
+
 boto3==1.7.84
+whitenoise==3.3.0
+
+# Api
+coreapi==2.3.3
+djangorestframework==3.8.2
+markdown==2.6.11
+pygments==2.2.0
+drf-writable-nested==0.4.1
+django-filter==2.0.0
